+++ tests/test_types.py
@@ -14,10 +14,7 @@
 # http://github.com/hearsaycorp/normalize
 #
 
-
-import six
 import decimal
-from builtins import object
 from datetime import date
 from datetime import datetime
 import sys
@@ -29,7 +26,6 @@ from normalize.property import Property
 from normalize.property import SafeProperty
 from normalize.property.types import *
 from normalize.subtype import subtype
-from future.utils import with_metaclass
 
 
 class TestTypeLibrary(unittest.TestCase):
@@ -59,24 +55,22 @@ class TestTypeLibrary(unittest.TestCase):
         self.assertEqual(demo.seq, 0)
         demo.name = "Foo Bar"
         self.assertEqual(demo.fullname, "Foo Bar")
-        self.assertIsInstance(demo.fullname, six.text_type)
+        self.assertIsInstance(demo.fullname, str)
 
         # FIXME: the actual errors returned in this situation are obtuse
-        with self.assertRaises(TypeError):
-            demo.name = 1
         with self.assertRaises(TypeError):
             demo.fullname = 123
 
         # test upgrade
         demo.fullname = str("foo")
-        self.assertIsInstance(demo.fullname, six.text_type)
+        self.assertIsInstance(demo.fullname, str)
 
         # no downgrade is attempted (or desirable tbh)
         demo.name = u"Bob"
-        self.assertIsInstance(demo.name, six.text_type)
+        self.assertIsInstance(demo.name, str)
 
         demo.num = "123"
-        self.assertIsInstance(demo.num, six.integer_types)
+        self.assertIsInstance(demo.num, int)
         demo.num = "123.0"
         self.assertIsInstance(demo.num, float)
         demo.num = "nan"
@@ -142,7 +136,6 @@ class TestSubTypes(unittest.TestCase):
     """Proof of concept test for coercing between sub-types of real types.
     """
     def test_sub_types(self):
-        long_type = six.integer_types[-1]
         NaturalNumber = subtype(
             of=int,
             name="NaturalNumber",
@@ -155,7 +148,7 @@ class TestSubTypes(unittest.TestCase):
         self.assertEqual(str(NaturalNumber), "<subtype NaturalNumber of int>")
 
         BigNaturalNumber = subtype(
-            of=long_type,
+            of=int,
             name="BigNaturalNumber",
             where=lambda i: i > 0,
         )
@@ -163,10 +156,7 @@ class TestSubTypes(unittest.TestCase):
         class NaturalBornObject(Record):
             count = Property(
                 isa=(NaturalNumber, BigNaturalNumber),
-                coerce=lambda x: (
-                    abs(int(x)) if abs(long_type(x)) < sys.maxsize else
-                    abs(long_type(x))
-                ),
+                coerce=lambda x: abs(int(x)),
                 check=lambda N: N > 0,
             )
 
@@ -175,7 +165,7 @@ class TestSubTypes(unittest.TestCase):
         nbo.count = "256"
         self.assertEqual(nbo.count, 256)
         nbo.count = 1.832e19
-        self.assertEqual(nbo.count, long_type(18320000000000000000))
+        self.assertEqual(nbo.count, int(18320000000000000000))
         # type matches, but subtype doesn't
         nbo.count = -10
         self.assertEqual(nbo.count, 10)
@@ -227,7 +217,7 @@ class TestSubTypes(unittest.TestCase):
     def test_subtype_abstract(self):
         import abc
 
-        class AbstractClass(with_metaclass(abc.ABCMeta, object)):
+        class AbstractClass(metaclass=abc.ABCMeta):
             @abc.abstractmethod
             def define_me(self):
                 pass